Moved navigation to its own file
